# Optimize the tidy task