--- a/tasks/go.py
+++ b/tasks/go.py
@@ -410,17 +410,15 @@ def tidy_all(ctx):
 
 
 @task
-def tidy(ctx, only_modified_packages=False):
-    # TODO: if only_modified_packages then use `go mod tidy -diff` with golang 1.23, probably in a separate check-mod-tidy task
-    # https://github.com/golang/go/issues/27005
-    from tasks import get_modified_packages
-
-    # TODO: Also include packages that import them
-    modules = get_modified_packages(ctx) if only_modified_packages else DEFAULT_MODULES.values()
-
-    for mod in modules:
+def tidy(ctx):
+    # Note: It's currently faster to tidy everything than looking for exactly what we should tidy
+    promises = []
+    for mod in DEFAULT_MODULES.values():
         with ctx.cd(mod.full_path()):
-            ctx.run("go mod tidy")
+            promises.append(ctx.run("go mod tidy", asynchronous=True))
+
+    for promise in promises:
+        promise.join()
 
 
 @task