<!--
* New contributors are highly encouraged to read our
  [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
-->
### What does this PR do?

<!--
* A brief description of the change being made with this pull request.
* If the description here cannot be expressed in a succinct form, consider
  opening multiple pull requests instead of a single one.
-->

Optimize the tidy task

### Motivation

<!--
* What inspired you to submit this pull request?
* Link any related GitHub issues or PRs here.
-->

- The current `--only-modified-packages` does not work well because it does not go through the modules that depend on the code you modified
- Doing this research takes ~7 seconds on my machine (after optimizations, otherwise it's 12 seconds)
- Running the tidy on everything after this optimization takes ~3 seconds on my machine, so it's not worth it to compute the exact modules to tidy

Tidying everything without optimization:

```
❯ time deva tidy
deva tidy 6.426 total
```

Tidying everything + optimization:

```
❯ time deva tidy
deva tidy 2.932 total
```

Note: The longest file to tidy takes 1.5 seconds on its own

### Additional Notes

<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->

https://datadoghq.atlassian.net/browse/ADXT-298

### Possible Drawbacks / Trade-offs

<!--
* What are the possible side-effects or negative impacts of the code change?
-->

### Describe how to test/QA your changes

<!--
* Write here in detail or link to detailed instructions on how this change can
  be tested/QAd/validated, including any environment setup.
-->